test: Clean up envd integration tests
This commit cleans up the environmentd integration tests by doing the
following:
  - Remove logging from all the tests. The logging has gotten extremely
  noisy and is not helpful for debugging test failures. However it is
  useful for production debugging, so we cannot remove them completely.
  - Move utility methods from the sql module to the util module.
